+++ CHANGELOG.md
@@ -3,6 +3,12 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 
+## [1.1.1] - 2020-05-13
+### Added
+Support for debug logging
+
+Debug logger writes to debug.log with -d flag set
+
 ## [1.1.0] - 2020-05-13
 ### Added
 Modal window support
@@ -19,5 +25,6 @@ Reworked view instantiation, code is much cleaner!
 ### Added
 Search functionality with vim-syntax
 
+[1.1.0]: https://github.com/kanbara/lisniks/compare/v1.1.0...v1.1.1
 [1.1.0]: https://github.com/kanbara/lisniks/compare/v1.0.0...v1.1.0
 [1.0.0]: https://github.com/kanbara/lisniks/releases/tag/v1.0.0